Task 1,

+++ 0x15-file_io/1-create_file.c
@@ -9,13 +9,13 @@
  */
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t o, w;
+	int o, w;
 	int len;
 
 	if (filename == NULL)
 		return (-1);
 
-	if (text_content !=NULL)
+	if (text_content != NULL)
 	{
 		for (len = 0; text_content[len];)
 			len++;
